+++ announcement-processor-parser/src/main/java/parser/scrapper/AnnouncementParser.java
@@ -1,5 +1,8 @@
-package parser;
+package parser.scrapper;
+
+import parser.Announcement;
 
 public abstract class AnnouncementParser implements SinglePageParser {
     public abstract Announcement parsePage(String url);
+    abstract String getPageContent(String url);
 }
